☀️ 20% – Back out 30 files of evidence-free retype edits
`1c44782f6` kept whatever a greedy sweep over seven return types happened to
move, on owners hundreds of halfwords from matching, and said so in its own
message: "No new bytes." That is not a cheap improvement, it is a false claim
written into the corpus. One of its hunks flips four callees from `void` to
`s32` because the positional score dropped by two; nothing establishes that any
of them returns anything.

Positional halfword counts are only comparable between candidates of identical
size, and are close to meaningless at that distance -- a rule this repository's
own ledger states and which the commit quoted before ignoring.

30 of the 34 files are restored. Three are left because the pool-word commit
later edited them on real evidence, and one is now exact.

What stays is what a fact produced: the return-type rule where it CLOSED an
owner, which is verified by bytes, and the pool-word rule, which encodes a
stated law rather than a search. The build still reproduces the ROM
byte-identically.

+++ semantic/resource_39c_c_020018f8.c
@@ -27,7 +27,7 @@ typedef struct ParticleOptions {
 
 extern void Func_02007316(s32 left, s32 top, s32 right, s32 bottom,
                           s32 style, s32 enabled);
-extern u8 Func_020072ce(void);
+extern u32 Func_020072ce(void);
 extern void Func_02001aa6(s32 x, s32 y, s32 z, s32 kind,
                           s32 velocity, s32 unused, s32 duration,
                           ParticleOptions *options);
